Add sort-operator-numeric
It's too complicated to support `:sort n`-style numeric sort char-wise,
but very simple and useful to support it line-wise. So do a basic
version for char-wise that sorts two numbers as numbers.

Change the sort to be stable since characters after a number make sorted
objects not uniquely identified by the sort key (the number).

Using <Plug>(sort-operator-numeric)ib, we get:

    (bb, aa, 45, cc, 54, 5, 39, 40aa, -10)
sorted to:
    (-10, 5, 39, 40aa, 45, 54, aa, bb, cc)
sorted to:
    (-10, 5, 39, 40aa, 45, 54, aa, bb, cc)

The sort is stable, sorts both strings and numbers, and positions
strings at the end.

`:sort n` would give:
    bb, aa, cc, -10, 5, 39, 40aa, 45, 54,

So the number section is identical, but the string section is at the
other end and sort-operator-numeric sorts it.

In contrast, using <Plug>(sort-operator)ib, we get:

    (bb, aa, 45, cc, 54, 5, 39, 40aa, -10)
sorted to:
    (-10, 39, 40aa, 45, 5, 54, aa, bb, cc)
sorted to:
    (-10, 39, 40aa, 45, 5, 54, aa, bb, cc)

Identical results to `:sort`.